fix: compute the budgeted-render pixel cap in f64 to avoid u64 wrap bypass
render_page_pdfium_budgeted bounded memory by estimating the max-DPI
pixel count as a u64 product of the two device-pixel dimensions. When
each factor is >= 2^32 the product wraps modulo 2^64 to a small value
that satisfies pixels_at_max <= max_pixels, so the DPI-reduction branch
is skipped and the page renders at full DPI -- the exact OOM the budget
exists to prevent.

Compute the estimate in f64 and compare against max_pixels as f64. An
overflowing product saturates toward inf and compares correctly, so the
cap can no longer be bypassed by an adversarial /MediaBox.

+++ src/pdf.rs
@@ -123,9 +123,9 @@ pub(crate) fn budgeted_render_dpi(
     max_pixels: u64,
 ) -> (u32, bool) {
     let scale_at_max = max_dpi as f64 / 72.0;
-    let pixels_at_max = (width_pts * scale_at_max) as u64 * (height_pts * scale_at_max) as u64;
+    let pixels_at_max = (width_pts * scale_at_max) * (height_pts * scale_at_max);
 
-    if pixels_at_max <= max_pixels {
+    if pixels_at_max <= max_pixels as f64 {
         (max_dpi, false)
     } else {
         // scale = sqrt(max_pixels / (w_pts * h_pts)), then dpi = scale * 72
